requirements.txt
fastapi[all]
uvicorn[standard]
pyjwt
httpx
pybit
binance-futures-connector